% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsdepot-package.R
\docType{package}
\name{plsdepot-package}
\alias{plsdepot-package}
\alias{plsdepot}
\title{plsdepot: Partial Least Squares (PLS) Data Analysis Methods}
\description{
Different methods for PLS analysis of one or two data tables such as Tucker's Inter-Battery, NIPALS, SIMPLS, SIMPLS-CA, PLS Regression, and PLS Canonical Analysis. The main reference for this software is the awesome book (in French) 'La Regression PLS: Theorie et Pratique' by Michel Tenenhaus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gastonstat/plsdepot/}
  \item Report bugs at \url{https://github.com/gastonstat/plsdepot/issues/}
}

}
\author{
This package was written by Gaston Sanchez. 
Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>
}
\keyword{internal}
