% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIP.R
\name{VIP_plsreg}
\alias{VIP_plsreg}
\title{Variable Influence on Projection (VIP)}
\usage{
VIP_plsreg(object)
}
\arguments{
\item{object}{\code{plsreg1} or \code{plsreg2} object.}
}
\description{
Variable Influence on Projection (VIP)
}
\references{
[1]. http://mevik.net/work/software/VIP.R
}
