% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plsreg1}
\alias{predict.plsreg1}
\title{Predict method for Partial Least Squares Regression 1 (plsreg1)}
\usage{
\method{predict}{plsreg1}(object, newdata, ...)
}
\arguments{
\item{object}{plsreg1 object.}

\item{newdata}{An data frame in which to look for variables with which to predict.}

\item{...}{other parameters will be ignore.}
}
\description{
Predict method for Partial Least Squares Regression 1 (plsreg1)
}
